chore: rename waterfall files (phase 3)
